agent/requirements.txt
langgraph
ipython
fastapi
uvicorn[standard]
pydantic
requests
psycopg[binary]
kubernetes
pyyaml
openai